Move operation and chain config dsl methods into own module
